Updating deployment diagram
Added a CA component to the StrongTNC device, which will be used by any
other components requiring a certificate for given connections. Note:
No detailed connections between the CA and other components are
depicted in the diagram.

Also indicated that the REST API to StrongTNC is preferred over the
optional direct SQL connection.